User stories implmented.

+++ app/controllers/questions_controller.rb
@@ -9,7 +9,7 @@ class QuestionsController < ApplicationController
   end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.new(question_params)
 
     if @question.save
       redirect_to @question, notice: 'Your question successfully created.'
@@ -19,6 +19,7 @@ class QuestionsController < ApplicationController
   end
 
   def show
+    @answers = question.answers.order(:id)
   end
 
   def edit
@@ -33,8 +34,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if question.author?(current_user)
+      question.destroy
+      redirect_to questions_path, notice: 'Question has been deleted.'
+    else
+      render :show
+    end
   end
 
   private
